src/cgi-detector-service/forensics/jpeg_dimples.py
import numpy as np
from PIL import Image
import io

def detect_jpeg_dimples(image_bytes: bytes) -> float:
    """
    Detects JPEG dimples artifacts in an image.

    JPEG dimples are a specific type of periodic artifact that can appear in JPEG-compressed images,
    serving as a valuable tool in photo forensics to detect image manipulation. These "dimples"
    manifest as a single darker or brighter pixel within each 8x8 pixel block of an image.

    Args:
        image_bytes: The raw bytes of the image.

    Returns:
        A score between 0.0 and 1.0, where a higher score indicates a higher
        probability of manipulation based on JPEG dimple analysis.
    """
    try:
        # Placeholder: In a real implementation, this would involve
        # decompressing the JPEG image to analyze DCT coefficients
        # and look for specific patterns indicative of dimples.
        # This is a complex operation that requires a deeper dive into JPEG
        # compression standards and potentially specialized libraries.
        image = Image.open(io.BytesIO(image_bytes)).convert('RGB')
        _ = np.array(image, dtype=np.uint8) # Just to show image is loaded
        
        # For now, return a placeholder score.
        # The actual implementation would analyze the 8x8 blocks for dimples.
        return 0.0
    except Exception:
        return 0.0
